refactor(api): _is_local_endpoint und LOCAL_NO_AUTH_API_KEY in ein neutrales Modul ziehen
Moved the `_is_local_endpoint` function, `_LOCAL_HOSTS` whitelist, and `LOCAL_NO_AUTH_API_KEY` constant from `simulation_prepare.py` to a newly created neutral utility module `backend/app/utils/endpoints.py` to fix their incorrect module residence. Renamed them to remove the leading underscore, and updated all source and test imports, keeping the semantic logic identical to preserve subdomain smuggling protection.

Fixes #800
